pyproject.toml
[project]
name = "indian-constitution-rag-chatbot"
version = "0.1.0"
description = "A RAG-based chatbot over the Indian Constitution using LangChain + OpenAI"
requires-python = ">=3.9"

dependencies = [
    "pandas",
    "openai",
    "langchain",
    "faiss-cpu",
    "tiktoken",
    "streamlit",
    "python-dotenv"
]
